# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Resource Manager API
  description: Resource Manager application for managing shared resources such as work rooms which contain desks. Employees, students of an institution can book them for certain time ranges.
  
  version: 1.0.0
...
paths:
...
  /rooms:
    get:
      tags:
        - admin
      summary: Retrieve info about all rooms available
      description: The admin can access the info about all the rooms on the app
      operationId: getRooms
      responses:
        '200':
          description: Succesfull retrieval of room info
        '400':
          description: Bad request
        '404':
          description: No rooms found
    post:
      tags:
        - admin
      summary: Create a room
      description: The admin can create rooms that contain desks to be booked by users
      operationId: createRoom
      responses:
        '200':
          description: Succesful creation of a room
        '400':
          description: Bad request
  /rooms/{room_number}:
    get:
      tags:
        - admin
      summary: Retrieve info about a room 
      description: The admin can retrieve info about a single room
      operationId: getRoom
      parameters:
        - name: room_number
          in: path
          description: The number of the room to be retrieved
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Succesful retrieval of room info
        '400':
          description: Bad request
        '404':
          description: Room not found
    post:
      tags:
        - admin
      summary: Update a room
      description: The admin can update a room's info
      operationId: updateRoom
      parameters:
        - name: room_number
          in: path
          description: The number of the room to be retrieved
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Succesful update of a room
        '400':
          description: Bad request
        '404':
          description: Room not found
  /rooms/{room_id}/desks:
    get:
      tags:
        - admin
      summary: Get desk's from a room info
      description: The admin can get the info of alll the desks in a room
      operationId: getRoomDesks
      parameters:
        - name: room_id
          in: path
          description: The id of the room to retrieve desk info from
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Succesful retrieval of a room's desks
        '400':
          description: Bad request
        '404':
          description: No desks found
  /desks:
    get:
      tags:
          - admin
      summary: Get info about all desks
      description: The admin can get the info of all the desks available
      operationId: getDesks
      responses:
        '200':
          description: Succesful retrieval of a desks' info
        '400':
          description: Bad request
        '404':
          description: No desks found
    post:
      tags:
          - admin
      summary: Create a desk
      description: The admin can create a desk for users to book
      operationId: createDesk
      responses:
        '200':
          description: Succesful creation of a desk
        '400':
          description: Bad request
  /desks/{desk_id}:
    get:
        tags:
            - admin
        summary: Get info about a desk
        description: The admin can get the info of a desk
        operationId: getDesk
        parameters:
        - name: desk_id
          in: path
          description: The id of the desk to get info from
          required: true
          schema:
            type: integer
            format: int64
        responses:
          '200':
            description: Succesful retrieval of a desk's info
          '400':
            description: Bad request
          '404':
            description: No desk found
    post:
        tags:
            - admin
        summary: Update info about a desk
        description: The admin can update the info of a desk
        operationId: updateDesk
        parameters:
        - name: desk_id
          in: path
          description: The id of the desk to update its info
          required: true
          schema:
            type: integer
            format: int64
        responses:
          '200':
            description: Succesful update of a desk's info
          '400':
            description: Bad request
          '404':
            description: No desk found
    
    delete:
      tags:
          - admin
      summary: Delete a desk
      description: The admin can delete a desk
      operationId: deleteDesk
      parameters:
      - name: desk_id
        in: path
        description: The id of the desk to delete 
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Succesful deletion of a desk
        '400':
          description: Bad request
        '404':
          description: No desk found
...